+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `Password::new()` renamed to `Password::generate()`.
+- `ValidatedConfig::generate()` renamed to `ValidatedConfig::generate_passwords()`.
+
 ## [1.1.4] - 2022-10-01
 <!--BEGIN=1.1.4-->
 Last 1.x release before refactor. Updated all dependencies.